internal/handler: merge into model
That package was too small, and its name was not obvious. I tried renaming
it and then realised it was simpler to just put its code into model.

The moved code is basically the default implementation of the callbacks and
it does not make much sense to separate interface anche implementation.

Late addition to https://github.com/ooni/probe-engine/issues/746.
