+++ experiment/example/example_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/apex/log"
 	"github.com/ooni/probe-engine/experiment/example"
-	"github.com/ooni/probe-engine/internal/handler"
 	"github.com/ooni/probe-engine/internal/mockable"
 	"github.com/ooni/probe-engine/model"
 )
@@ -20,7 +19,7 @@ func TestIntegrationSuccess(t *testing.T) {
 	sess := &mockable.ExperimentSession{
 		MockableLogger: log.Log,
 	}
-	callbacks := handler.NewPrinterCallbacks(sess.Logger())
+	callbacks := model.NewPrinterCallbacks(sess.Logger())
 	err := m.Run(ctx, sess, new(model.Measurement), callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -36,7 +35,7 @@ func TestIntegrationFailure(t *testing.T) {
 	sess := &mockable.ExperimentSession{
 		MockableLogger: log.Log,
 	}
-	callbacks := handler.NewPrinterCallbacks(sess.Logger())
+	callbacks := model.NewPrinterCallbacks(sess.Logger())
 	err := m.Run(ctx, sess, new(model.Measurement), callbacks)
 	if err == nil {
 		t.Fatal("expected an error here")
