+++ experiment/hhfm/hhfm_test.go
@@ -16,7 +16,6 @@ import (
 	engine "github.com/ooni/probe-engine"
 	"github.com/ooni/probe-engine/experiment/hhfm"
 	"github.com/ooni/probe-engine/experiment/urlgetter"
-	"github.com/ooni/probe-engine/internal/handler"
 	"github.com/ooni/probe-engine/internal/mockable"
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/archival"
@@ -39,7 +38,7 @@ func TestIntegrationSuccess(t *testing.T) {
 	// we need a real session because we need the tcp-echo helper
 	sess := newsession(t)
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -143,7 +142,7 @@ func TestIntegrationCancelledContext(t *testing.T) {
 	// we need a real session because we need the tcp-echo helper
 	sess := newsession(t)
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -245,7 +244,7 @@ func TestNoHelpers(t *testing.T) {
 	ctx := context.Background()
 	sess := &mockable.ExperimentSession{}
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, hhfm.ErrNoAvailableTestHelpers) {
 		t.Fatal("not the error we expected")
@@ -295,7 +294,7 @@ func TestNoActualHelpersInList(t *testing.T) {
 		},
 	}
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, hhfm.ErrNoAvailableTestHelpers) {
 		t.Fatal("not the error we expected")
@@ -348,7 +347,7 @@ func TestWrongTestHelperType(t *testing.T) {
 		},
 	}
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, hhfm.ErrInvalidHelperType) {
 		t.Fatal("not the error we expected")
@@ -401,7 +400,7 @@ func TestNewRequestFailure(t *testing.T) {
 		},
 	}
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err == nil || !strings.HasSuffix(err.Error(), "invalid control character in URL") {
 		t.Fatal("not the error we expected")
@@ -458,7 +457,7 @@ func TestInvalidJSONBody(t *testing.T) {
 		},
 	}
 	measurement := new(model.Measurement)
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -506,7 +505,7 @@ func TestTransactStatusCodeFailure(t *testing.T) {
 		StatusCode: 500,
 	}}
 	resp, body, err := hhfm.Transact(txp, &http.Request{},
-		handler.NewPrinterCallbacks(log.Log))
+		model.NewPrinterCallbacks(log.Log))
 	if !errors.Is(err, urlgetter.ErrHTTPRequestFailed) {
 		t.Fatal("not the error we expected")
 	}
@@ -525,7 +524,7 @@ func TestTransactCannotReadBody(t *testing.T) {
 		StatusCode: 200,
 	}}
 	resp, body, err := hhfm.Transact(txp, &http.Request{},
-		handler.NewPrinterCallbacks(log.Log))
+		model.NewPrinterCallbacks(log.Log))
 	if !errors.Is(err, expected) {
 		t.Fatal("not the error we expected")
 	}
