+++ experiment/webconnectivity/webconnectivity_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	engine "github.com/ooni/probe-engine"
 	"github.com/ooni/probe-engine/experiment/webconnectivity"
-	"github.com/ooni/probe-engine/internal/handler"
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/archival"
 	"github.com/ooni/probe-engine/netx/errorx"
@@ -32,7 +31,7 @@ func TestIntegrationSuccess(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, true)
 	measurement := &model.Measurement{Input: "http://www.example.com"}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -60,7 +59,7 @@ func TestMeasureWithCancelledContext(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, true)
 	measurement := &model.Measurement{Input: "http://www.example.com"}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if err != nil {
 		t.Fatal(err)
@@ -88,7 +87,7 @@ func TestMeasureWithNoInput(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, true)
 	measurement := &model.Measurement{Input: ""}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, webconnectivity.ErrNoInput) {
 		t.Fatal(err)
@@ -116,7 +115,7 @@ func TestMeasureWithInputNotBeingAnURL(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, true)
 	measurement := &model.Measurement{Input: "\t\t\t\t\t\t"}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, webconnectivity.ErrInputIsNotAnURL) {
 		t.Fatal(err)
@@ -144,7 +143,7 @@ func TestMeasureWithUnsupportedInput(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, true)
 	measurement := &model.Measurement{Input: "dnslookup://example.com"}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, webconnectivity.ErrUnsupportedInput) {
 		t.Fatal(err)
@@ -172,7 +171,7 @@ func TestMeasureWithNoAvailableTestHelpers(t *testing.T) {
 	// we need a real session because we need the web-connectivity helper
 	sess := newsession(t, false)
 	measurement := &model.Measurement{Input: "https://www.example.com"}
-	callbacks := handler.NewPrinterCallbacks(log.Log)
+	callbacks := model.NewPrinterCallbacks(log.Log)
 	err := measurer.Run(ctx, sess, measurement, callbacks)
 	if !errors.Is(err, webconnectivity.ErrNoAvailableTestHelpers) {
 		t.Fatal(err)
