+++ model/experiment.go
@@ -5,6 +5,8 @@ import (
 	"net/http"
 	"net/url"
 	"time"
+
+	"github.com/ooni/probe-engine/internal/humanizex"
 )
 
 // ExperimentOrchestraClient is the experiment's view of
@@ -50,6 +52,29 @@ type ExperimentCallbacks interface {
 
 	// OnProgress provides information about an experiment progress.
 	OnProgress(percentage float64, message string)
+}
+
+// PrinterCallbacks is the default event handler
+type PrinterCallbacks struct {
+	Logger
+}
+
+// NewPrinterCallbacks returns a new default callback handler
+func NewPrinterCallbacks(logger Logger) PrinterCallbacks {
+	return PrinterCallbacks{Logger: logger}
+}
+
+// OnDataUsage provides information about data usage.
+func (d PrinterCallbacks) OnDataUsage(dloadKiB, uploadKiB float64) {
+	d.Logger.Infof("experiment: recv %s, sent %s",
+		humanizex.SI(dloadKiB*1024, "byte"),
+		humanizex.SI(uploadKiB*1024, "byte"),
+	)
+}
+
+// OnProgress provides information about an experiment progress.
+func (d PrinterCallbacks) OnProgress(percentage float64, message string) {
+	d.Logger.Infof("[%5.1f%%] %s", percentage*100, message)
 }
 
 // ExperimentMeasurer is the interface that allows to run a
